marketbard/requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.7
dotenv==0.9.9
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
kafka==1.3.5
kafka-python-ng==2.2.3
numpy==2.3.2
packaging==25.0
pycparser==2.22
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
threadpoolctl==3.6.0
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
urllib3==2.5.0
